api/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
python-dotenv==1.0.0
flask-login==0.6.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
psycopg2-binary==2.9.9
markupsafe==2.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
click==8.1.7
